Removed duplication and moved everything to one file
