+++ geoopt/samplers/__init__.py
@@ -1,3 +1,1 @@
-from .sgld import SGLD, RSGLD
-from .hmc import HMC, SGHMC
-from .rhmc import RHMC, SGRHMC
+from .samplers import RSGLD, RHMC, SGRHMC
