+++ geoopt/samplers/samplers.py
@@ -1,15 +1,89 @@
 import math
 import numpy as np
 import torch
+import torch.optim as optim
 
 from ..tensor import ManifoldParameter, ManifoldTensor
 from ..manifolds import Rn
 from ..optim.mixin import OptimMixin
-from .hmc import HMC, SGHMC
+
+
+class Sampler(optim.Optimizer):
+    def __init__(self, params, defaults):
+        super().__init__(params, defaults)
+        self.n_rejected = 0
+        self.steps = 0
+        self.burnin = True
+
+        self.log_probs = []
+        self.acceptance_probs = []
+    
+        
+    @property
+    def rejection_rate(self):
+        if self.steps > 0:
+            return self.n_rejected / self.steps
+        else:
+            return 0.0
+
+
+class RSGLD(OptimMixin, SGLD):
+    """Riemannian Stochastic Gradient Langevin Dynamics"""
+
+    def __init__(self, params, epsilon=1e-3):
+        defaults = dict(epsilon=epsilon)
+        super().__init__(params, defaults)
+    
+
+    def step(self, closure):
+        """Performs a single sampling step.
+
+        Arguments:
+            closure (callable, optional): A closure that reevaluates the model
+                and returns the log probability.
+        """
+        logp = closure()
+        logp.backward()
+
+        with torch.no_grad():
+            for group in self.param_groups:
+                for p in group['params']:
+                    if isinstance(p, (ManifoldParameter, ManifoldTensor)):
+                        manifold = p.manifold
+                    else:
+                        manifold = Rn()
+
+                    proju, retr = manifold.proju, manifold.retr
+                    epsilon = group['epsilon']
+
+                    n = torch.randn_like(p).mul_(math.sqrt(epsilon))
+                    r = proju(p.data, 0.5 * epsilon * p.grad + n)
+
+                    p.data.set_(retr(p.data, r, 1.))
+                    p.grad.data.zero_()
+
+        if not self.burnin:
+            self.steps += 1
+            self.log_probs.append(logp.item())
+
+
+    def stabilize(self):
+        for group in self.param_groups:
+            for p in group["params"]:
+                if not isinstance(p, (ManifoldParameter, ManifoldTensor)):
+                    continue
+
+                p.data.set_(p.manifold.projx(p.data))
 
 
 class RHMC(OptimMixin, HMC):
     """Riemannian Hamiltonian Monte-Carlo"""
+
+    def __init__(self, params, epsilon=1e-3, n_steps=1):
+        defaults = dict(epsilon=epsilon)
+        super(HMC, self).__init__(params, defaults)
+        self.n_steps = n_steps
+
 
     def step(self, closure):
         """Performs a single sampling step.
@@ -146,6 +220,12 @@ class RHMC(OptimMixin, HMC):
 
 class SGRHMC(SGHMC):
     """Stochastic Gradient Riemannian Hamiltonian Monte-Carlo"""
+    
+    def __init__(self, params, epsilon=1e-3, n_steps=1, alpha=0.1):
+        defaults = dict(epsilon=epsilon, alpha=alpha)
+        super(SGHMC, self).__init__(params, defaults)
+        self.n_steps = n_steps
+    
 
     def step(self, closure):
         """Performs a single sampling step.
